+++ src/main/kotlin/cloud/glitchdev/rfu/achievement/AchievementType.kt
@@ -3,5 +3,5 @@ package cloud.glitchdev.rfu.achievement
 enum class AchievementType {
     NORMAL, // Fully visible
     SECRET, // Title/Icon visible, description hidden
-    HIDDEN  // Invisible until completed
+    OBFUSCATED  // Obfuscated text until completed
 }