+++ src/main/kotlin/cloud/glitchdev/rfu/achievement/types/DyeAchievement.kt
@@ -1,4 +1,4 @@
-﻿package cloud.glitchdev.rfu.achievement.types
+package cloud.glitchdev.rfu.achievement.types
 
 import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import cloud.glitchdev.rfu.achievement.BaseAchievement
@@ -23,10 +23,8 @@ abstract class DyeAchievement(val dye: Dyes) : BaseAchievement() {
             }
         })
 
-        activeListeners.add(ContainerEvents.registerContainerOpenEvent { _, items ->
-            //~ if >=26.2 'screen' -> 'gui.screen()' {
-            if (mc.gui.screen()?.title?.string == "Dye Compendium") {
-            //~}
+        activeListeners.add(ContainerEvents.registerContainerOpenEvent { containerName, _, items ->
+            if (containerName == "Dye Compendium") {
                 val item = items.find { it.hoverName.toUnformattedString() == dye.dyeName }
                     ?: return@registerContainerOpenEvent
                 val lore = item[DataComponents.LORE] ?: return@registerContainerOpenEvent