+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/ContainerEvents.kt
@@ -1,18 +1,23 @@
 package cloud.glitchdev.rfu.events.managers
 
+import cloud.glitchdev.rfu.RiccioFishingUtils.mc
 import cloud.glitchdev.rfu.events.AbstractEventManager
+import gg.essential.universal.utils.toUnformattedString
 import net.minecraft.world.item.ItemStack
 
-object ContainerEvents : AbstractEventManager<(containerId : Int, itens : List<ItemStack>) -> Unit, ContainerEvents.ContainerOpenEvent>() {
+object ContainerEvents : AbstractEventManager<(containerId: Int, itens: List<ItemStack>) -> Unit, ContainerEvents.ContainerOpenEvent>() {
     override val runTasks: (Int, List<ItemStack>) -> Unit = { containerId, itens ->
         safeExecution {
-            tasks.forEach { event -> event.callback(containerId, itens) }
+            //~ if >=26.2 'mc.screen' -> 'mc.gui.screen()' {
+            val containerName = mc.gui.screen()?.title?.toUnformattedString() ?: ""
+            //~}
+            tasks.forEach { event -> event.userCallback(containerName, containerId, itens) }
         }
     }
 
     fun registerContainerOpenEvent(
         priority: Int = 20,
-        callback: (containerId : Int, itens : List<ItemStack>) -> Unit
+        callback: (containerName: String, containerId: Int, itens: List<ItemStack>) -> Unit
     ): ContainerOpenEvent {
         return ContainerOpenEvent(priority, callback).register()
     }
@@ -20,8 +25,8 @@ object ContainerEvents : AbstractEventManager<(containerId : Int, itens : List<I
 
     class ContainerOpenEvent(
         priority: Int = 20,
-        callback: (containerId : Int, itens : List<ItemStack>) -> Unit
-    ) : ManagedTask<(containerId : Int, itens : List<ItemStack>) -> Unit, ContainerOpenEvent>(priority, callback) {
+        val userCallback: (containerName: String, containerId: Int, itens: List<ItemStack>) -> Unit
+    ) : ManagedTask<(containerId: Int, itens: List<ItemStack>) -> Unit, ContainerOpenEvent>(priority, { _, _ -> }) {
         override fun register() = submitTask(this)
         override fun unregister() = removeTask(this)
     }