+++ src/main/kotlin/cloud/glitchdev/rfu/feature/other/EmojiSuggestion.kt
@@ -8,5 +8,7 @@ class EmojiSuggestion(
     val trigger: String,
     val emoji: String
 ) : Suggestion(range, trigger) {
-    val displayText: String = "$emoji $trigger"
+    // This weird thing is to prevent it from being replaced by the feature, leading it to having the background of the wrong size
+    val displayText: String = "$emoji :§r${trigger.drop(1)
+    }"
 }