+++ CHANGELOG.md
@@ -1,9 +1,24 @@
-# v1.15.2 - Expressive & Empowered Patches Part 2
+# v1.16.0 - Torrhus Canyon
+
+### Features
+- Added a setting to toggle showing only recently caught trophies on the Trophy Pity Display.
+- Added a setting to toggle whether Enter completes emoji suggestions.
+- Added navigation to view different achievement tiers on staged achievements in /rfuachievements.
+- Added Torrhus Canyon to rfu party finder
+- Added Torrhus Canyon Sea Creatures
+- Added a hover message to expired cake alerts showing which cakes expired and a note on how to reset if they're incorrect.
 
 ### Fixes
-- Fixed an issue causing missing item textures when using the resourcepack feature
-- Made pity display work with 2/4 trophy hunter armor
+- Fixed command suggestion box width calculation conflict with SkyHanni for emojis.
+- Fixed trophy fish and frog pity display counting as one when caught with a multiplier (e.g. Geometric).
+- Fixed resource pack feature not updating the resourcepacks and saving because of file locks.
+- Added more validations before registering the hud to prevent a somewhat rare crash.
 
 ### Changes
-- Made the resourcepack feature off by default
-  - Also will disable it for any player who had it on, so if you wish it back on, go turn it on manually
+- Improved Search on the settings so it doesn't need to open the category
+- Set the minimum required number of chars to display the emoji suggestions to 2 and made it configurable on Other Settings.
+- Improved Pity Display to also update upon catching a fish instead of being on a 1s timer
+- Added more validation on party finder
+- Changed Atoll color on party finder to a cherry blossom color
+- Made full inventory alert off by default
+- Made hidden achievements visible on the achievement window but obfuscated.